+++ src/controllers/StripeController.php
@@ -76,7 +76,7 @@ class StripeController extends BaseController
     }
 
     /**
-     * @return null|\yii\web\Response
+     * @return \yii\web\Response|null
      * @throws \yii\web\BadRequestHttpException
      */
     public function actionCancelSubscription()
@@ -84,8 +84,53 @@ class StripeController extends BaseController
         $this->requirePostRequest();
         $request = Craft::$app->getRequest();
         $subscriptionId = $request->getRequiredBodyParam('subscriptionId');
+        $cancelAtPeriodEnd = null;
+        if (isset($_POST['cancelAtPeriodEnd'])){
+            $cancelAtPeriodEnd = filter_var ($_POST['cancelAtPeriodEnd'], FILTER_VALIDATE_BOOLEAN);
+        }
 
-        $result = StripePlugin::$app->subscriptions->cancelStripeSubscription($subscriptionId);
+        if (is_null($cancelAtPeriodEnd)){
+            $settings = StripePlugin::$app->settings->getSettings();
+            $cancelAtPeriodEnd = $settings->cancelAtPeriodEnd;
+        }
+
+        $result = StripePlugin::$app->subscriptions->cancelStripeSubscription($subscriptionId, $cancelAtPeriodEnd);
+
+        if (!$result) {
+            if (Craft::$app->getRequest()->getAcceptsJson()) {
+                return $this->asJson([
+                    'success' => false,
+                ]);
+            }
+
+            Craft::$app->getUrlManager()->setRouteParams([
+                    'success' => false
+                ]
+            );
+
+            return null;
+        }
+
+        if (Craft::$app->getRequest()->getAcceptsJson()) {
+            return $this->asJson([
+                'success' => true
+            ]);
+        }
+
+        return $this->redirectToPostedUrl();
+    }
+
+    /**
+     * @return \yii\web\Response|null
+     * @throws \yii\web\BadRequestHttpException
+     */
+    public function actionReactivateSubscription()
+    {
+        $this->requirePostRequest();
+        $request = Craft::$app->getRequest();
+        $subscriptionId = $request->getRequiredBodyParam('subscriptionId');
+
+        $result = StripePlugin::$app->subscriptions->reactivateStripeSubscription($subscriptionId);
 
         if (!$result) {
             if (Craft::$app->getRequest()->getAcceptsJson()) {