+++ src/elements/actions/SetStatus.php
@@ -28,7 +28,10 @@ class SetStatus extends ElementAction
 
     /**
      * @inheritdoc
-     * @throws \yii\base\Exception
+     * @return string
+     * @throws \Twig\Error\LoaderError
+     * @throws \Twig\Error\RuntimeError
+     * @throws \Twig\Error\SyntaxError
      */
     public function getTriggerHtml(): string
     {